it works

+++ MVC/SurfsUpWebApp/SurfsUpWebApp/Models/Booking.cs
@@ -11,10 +11,10 @@
         public TimeOnly RentHours { get; set; }
         public DateTime RentReturn { get; set; }
         public string Remarks { get; set; }
-        public int Price  { get; set; }
+        public float Price  { get; set; }
         public string SelectedPrice {  get; set; }
         public int SurfboardAmount { get; set; }
-        public Booking(int bookingId, string customerName, string customerEmail, string customerPhone, string selectedSurfboard, DateTime rentPeriod, TimeOnly rentHours, DateTime rentReturn, string remarks, int price, string selectedPrice, int surfboardAmount )
+        public Booking(int bookingId, string customerName, string customerEmail, string customerPhone, string selectedSurfboard, DateTime rentPeriod, TimeOnly rentHours, DateTime rentReturn, string remarks, float price, string selectedPrice, int surfboardAmount )
         {
             BookingId = bookingId;
             CustomerName = customerName;
